+++ src/EtcdTerminal/Configuration/IConnectionConfigRepository.cs
@@ -3,6 +3,7 @@ namespace EtcdTerminal.Configuration;
 public interface IConnectionConfigRepository
 {
 	IReadOnlyList<EtcdConnectionConfig> LoadInstances();
+	IReadOnlyList<string> TakeDecryptFailures();
 	void AddInstance(EtcdConnectionConfig config);
 	void UpdateInstance(string originalName, EtcdConnectionConfig config);
 	void RemoveInstance(string name);